+++ cmd/osv-scanner/scan/source/command_test.go
@@ -159,6 +159,37 @@ func TestCommand(t *testing.T) {
 			Args: []string{"", "source", "--recursive", "--no-ignore", "./testdata/locks-gitignore"},
 			Exit: 0,
 		},
+		// experimental exclude flag tests
+		{
+			Name: "exclude_with_exact_directory_name",
+			Args: []string{"", "source", "--recursive", "--experimental-exclude=nested", "./testdata/locks-one-with-nested"},
+			Exit: 0,
+		},
+		{
+			Name: "exclude_with_glob_pattern",
+			Args: []string{"", "source", "--recursive", "--experimental-exclude=g:**/nested/**", "./testdata/locks-one-with-nested"},
+			Exit: 0,
+		},
+		{
+			Name: "exclude_with_regex_pattern",
+			Args: []string{"", "source", "--recursive", "--experimental-exclude=r:/nested$", "./testdata/locks-one-with-nested"},
+			Exit: 0,
+		},
+		{
+			Name: "exclude_with_invalid_regex_returns_error",
+			Args: []string{"", "source", "--experimental-exclude=r:[invalid", "./testdata/locks-many"},
+			Exit: 127,
+		},
+		{
+			Name: "exclude_with_multiple_exact_directories",
+			Args: []string{"", "source", "--recursive", "--experimental-exclude=nested", "--experimental-exclude=other", "./testdata/locks-one-with-nested"},
+			Exit: 0,
+		},
+		{
+			Name: "exclude_with_multiple_pattern_types",
+			Args: []string{"", "source", "--recursive", "--experimental-exclude=nested", "--experimental-exclude=g:**/vendor/**", "--experimental-exclude=r:\\.cache$", "./testdata/locks-one-with-nested"},
+			Exit: 0,
+		},
 		{
 			Name: "json output",
 			Args: []string{"", "source", "--format", "json", "./testdata/locks-many/composer.lock"},
@@ -1302,6 +1333,11 @@ func TestCommand_Transitive(t *testing.T) {
 		{
 			Name: "transitive_requirements_enricher_requires_enabled_requirements_extractor",
 			Args: []string{"", "source", "--experimental-disable-plugins=python/requirements", "./testdata/locks-requirements/requirements-transitive.txt"},
+			Exit: 128,
+		},
+		{
+			Name: "transitive_pomxml_enricher_requires_enabled_pomxml_extractor",
+			Args: []string{"", "source", "--experimental-disable-plugins=java/pomxml", "./testdata/maven-transitive/abc.xml"},
 			Exit: 128,
 		},
 	}