+++ internal/cachedregexp/regex.go
@@ -1,4 +1,4 @@
-// Package cachedregexp provides a cached version of regexp.MustCompile.
+// Package cachedregexp provides a cached version of regexp.MustCompile and regexp.Compile.
 package cachedregexp
 
 import (
@@ -15,4 +15,22 @@ func MustCompile(exp string) *regexp.Regexp {
 	}
 
 	return compiled.(*regexp.Regexp)
+}
+
+// Compile returns a compiled regexp or an error if the pattern is invalid.
+// Results are cached for performance.
+func Compile(exp string) (*regexp.Regexp, error) {
+	compiled, ok := cache.Load(exp)
+	if ok {
+		return compiled.(*regexp.Regexp), nil
+	}
+
+	r, err := regexp.Compile(exp)
+	if err != nil {
+		return nil, err
+	}
+
+	cache.LoadOrStore(exp, r)
+
+	return r, nil
 }