+++ pkg/osvscanner/scan.go
@@ -15,9 +15,11 @@ import (
 	"github.com/google/osv-scalibr/enricher"
 	"github.com/google/osv-scalibr/enricher/packagedeprecation"
 	"github.com/google/osv-scalibr/enricher/reachability/java"
+	transitivedependencypomxml "github.com/google/osv-scalibr/enricher/transitivedependency/pomxml"
 	transitivedependencyrequirements "github.com/google/osv-scalibr/enricher/transitivedependency/requirements"
 	"github.com/google/osv-scalibr/extractor"
 	"github.com/google/osv-scalibr/extractor/filesystem"
+	"github.com/google/osv-scalibr/extractor/filesystem/language/java/pomxml"
 	"github.com/google/osv-scalibr/extractor/filesystem/language/python/requirements"
 	"github.com/google/osv-scalibr/extractor/filesystem/simplefileapi"
 	"github.com/google/osv-scalibr/fs"
@@ -26,7 +28,6 @@ import (
 	"github.com/google/osv-scalibr/plugin"
 	"github.com/google/osv-scanner/v2/internal/cmdlogger"
 	"github.com/google/osv-scanner/v2/internal/scalibrextract/filesystem/vendored"
-	"github.com/google/osv-scanner/v2/internal/scalibrextract/language/java/pomxmlenhanceable"
 	"github.com/google/osv-scanner/v2/internal/scalibrextract/vcs/gitcommitdirect"
 	"github.com/google/osv-scanner/v2/internal/scalibrextract/vcs/gitrepo"
 	"github.com/google/osv-scanner/v2/internal/scalibrplugin"
@@ -38,25 +39,6 @@ var ErrExtractorNotFound = errors.New("could not determine extractor suitable to
 
 func configurePlugins(plugins []plugin.Plugin, accessors ExternalAccessors, actions ScannerActions) {
 	for _, plug := range plugins {
-		if !actions.TransitiveScanning.Disabled {
-			err := pomxmlenhanceable.EnhanceIfPossible(plug, &cpb.PluginConfig{
-				UserAgent: actions.RequestUserAgent,
-				PluginSpecific: []*cpb.PluginSpecificConfig{
-					{
-						Config: &cpb.PluginSpecificConfig_PomXmlNet{
-							PomXmlNet: &cpb.POMXMLNetConfig{
-								UpstreamRegistry:    actions.TransitiveScanning.MavenRegistry,
-								DepsDevRequirements: !actions.TransitiveScanning.NativeDataSource,
-							},
-						},
-					},
-				},
-			})
-			if err != nil {
-				log.Errorf("Failed to enhance pomxml extractor: %v", err)
-			}
-		}
-
 		vendored.Configure(plug, vendored.Config{
 			// Only attempt to vendor check git directories if we are not skipping scanning root git directories
 			ScanGitDir: !actions.IncludeGitRoot,
@@ -68,6 +50,18 @@ func configurePlugins(plugins []plugin.Plugin, accessors ExternalAccessors, acti
 func isRequirementsExtractorEnabled(plugins []plugin.Plugin) bool {
 	for _, plug := range plugins {
 		_, ok := plug.(*requirements.Extractor)
+
+		if ok {
+			return true
+		}
+	}
+
+	return false
+}
+
+func isPomXMLExtractorEnabled(plugins []plugin.Plugin) bool {
+	for _, plug := range plugins {
+		_, ok := plug.(*pomxml.Extractor)
 
 		if ok {
 			return true
@@ -92,15 +86,39 @@ func getPlugins(defaultPlugins []string, accessors ExternalAccessors, actions Sc
 
 	plugins := scalibrplugin.Resolve(actions.PluginsEnabled, actions.PluginsDisabled)
 
-	// TODO: Use Enricher.RequiredPlugins to check this generically
-	if !actions.TransitiveScanning.Disabled && isRequirementsExtractorEnabled(plugins) {
-		p, err := transitivedependencyrequirements.New(&cpb.PluginConfig{
-			UserAgent: actions.RequestUserAgent,
-		})
-		if err != nil {
-			log.Errorf("Failed to make transitivedependencyrequirements enricher: %v", err)
-		} else {
-			plugins = append(plugins, p)
+	if !actions.TransitiveScanning.Disabled {
+		// TODO: Use Enricher.RequiredPlugins to check this generically
+		if isRequirementsExtractorEnabled(plugins) {
+			p, err := transitivedependencyrequirements.New(&cpb.PluginConfig{
+				UserAgent: actions.RequestUserAgent,
+			})
+			if err != nil {
+				log.Errorf("Failed to make transitivedependencyrequirements enricher: %v", err)
+			} else {
+				plugins = append(plugins, p)
+			}
+		}
+
+		// TODO: Use Enricher.RequiredPlugins to check this generically
+		if isPomXMLExtractorEnabled(plugins) {
+			p, err := transitivedependencypomxml.New(&cpb.PluginConfig{
+				UserAgent: actions.RequestUserAgent,
+				PluginSpecific: []*cpb.PluginSpecificConfig{
+					{
+						Config: &cpb.PluginSpecificConfig_PomXmlNet{
+							PomXmlNet: &cpb.POMXMLNetConfig{
+								UpstreamRegistry:    actions.TransitiveScanning.MavenRegistry,
+								DepsDevRequirements: !actions.TransitiveScanning.NativeDataSource,
+							},
+						},
+					},
+				},
+			})
+			if err != nil {
+				log.Errorf("Failed to make transitivedependencypomxml enricher: %v", err)
+			} else {
+				plugins = append(plugins, p)
+			}
 		}
 	}
 
@@ -235,6 +253,12 @@ SBOMLoop:
 	testlogger.BeginDirScanMarker()
 	osCapability := determineOS()
 
+	// Parse exclude patterns (supports exact names, glob, and regex)
+	excludePatterns, err := parseExcludePatterns(actions.ExcludePatterns)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse exclude patterns: %w", err)
+	}
+
 	// For each root, run scalibr's scan() once.
 	for root, paths := range rootMap {
 		capabilities := plugin.Capabilities{
@@ -254,14 +278,13 @@ SBOMLoop:
 			ScanRoots:             fs.RealFSScanRoots(root),
 			PathsToExtract:        paths,
 			IgnoreSubDirs:         !actions.Recursive,
-			DirsToSkip:            nil,
-			SkipDirRegex:          nil,
-			SkipDirGlob:           nil,
+			DirsToSkip:            excludePatterns.dirsToSkip,
+			SkipDirRegex:          excludePatterns.regexPattern,
+			SkipDirGlob:           excludePatterns.globPattern,
 			UseGitignore:          !actions.NoIgnore,
 			Stats:                 &statsCollector,
 			ReadSymlinks:          false,
 			MaxInodes:             0,
-			StoreAbsolutePath:     true,
 			PrintDurationAnalysis: false,
 			ErrorOnFSErrors:       false,
 			ExplicitPlugins:       true,
@@ -289,17 +312,32 @@ SBOMLoop:
 						// If there is more than 1 file error, write them on new lines
 						builder.WriteString("\n\t")
 					}
-					builder.WriteString(fmt.Sprintf("%s: %s", fileError.FilePath, fileError.ErrorMessage))
+					fmt.Fprintf(&builder, "%s: %s", fileError.FilePath, fileError.ErrorMessage)
 
 					// Check if the erroring file was a path specifically passed in (not a result of a file walk)
 					if slices.Contains(specificPaths, filepath.Join(root, fileError.FilePath)) {
 						criticalError = true
 					}
 				}
-				cmdlogger.Errorf("Error during extraction: (extracting as %s) %s", status.Name, builder.String())
+
+				msg := status.Status.FailureReason
+
+				if len(status.Status.FileErrors) > 0 {
+					msg = builder.String()
+				}
+
+				cmdlogger.Errorf("Error during extraction: (extracting as %s) %s", status.Name, msg)
 				if criticalError {
 					return nil, errors.New("extraction failed on specified lockfile")
 				}
+			}
+		}
+
+		// todo: ideally we'd have scalibr make package locations absolute this via StoreAbsolutePath,
+		//  but currently that can break plugins that don't support absolute paths, like the pomxml enricher
+		for _, pkg := range sr.Inventory.Packages {
+			for i, loc := range pkg.Locations {
+				pkg.Locations[i] = filepath.Join(root, loc)
 			}
 		}