+++ scripts/run_lints.sh
@@ -2,5 +2,5 @@
 
 set -ex
 
-export GOTOOLCHAIN="${GOTOOLCHAIN:-go1.25.6}"
-go run github.com/golangci/golangci-lint/v2/cmd/golangci-lint@v2.6.1 run ./... "$@"
+export GOTOOLCHAIN="${GOTOOLCHAIN:-go1.25.7}"
+go run github.com/golangci/golangci-lint/v2/cmd/golangci-lint@v2.8.0 run ./... "$@"